Expose Term, Renderer and RawMode for custom io.
The Editor struct has been renamed to GenericEditor and is now
parameterized over T: Term which permits user supplied instances of Term
to be used. Like this using rustyline in combination with other
libraries that rely on handling stdin and stdout themselves becomes
possible and in theory rustyline could even be used inside a gui
application.

To keep API changes minimal a type alias Editor is provided for
GenericEditor with tty::Terminal as Term.